refactor(engine): move the author-association vocabulary into the engine package
The gate-advisory TWIN carried a FIFTH copy of the OWNER/MEMBER/COLLABORATOR
predicate, and the engine is standalone -- it cannot import out of src/. So the
definition lives in @loopover/engine and src/github/author-association.ts
re-exports it, the same shape settings/pr-type-label.ts already uses for exactly
this reason. Both twins now read one definition.
